Recreated n-gram tuples as a new tuple type.
This was necessary in order to be able to pickle the data with a
definition of a named tuple that was defined in its own separate file
rather than in nonsense_detector.py.  This solved a portability issue
that I couldn't seem to solve before.  Unfortunately, it meant having
to redefine all the n-gram data values.