Moved definition of n-gram tuple type to separate file.
This was necessary in order to solve a problem of unpickling the data
file.  When the tuple type was left in the nonsense_detector.py file,
I couldn't get that file to define `is_nonsense()` because that
required unpickling the data, and that in turn failed when
nonsense_detector.py was imported by another file.  Putting the tuple
definition in a separate file made it possible for the n-gram data
file to be unpickled properly when nonsense_detector.py was imported
by another file.